Close dead-stack key residue in Ed25519 and Ascon; canonical FROST shares; strict-warning lanes at source
- Ed25519: the residue probe now searches the comb's 21-bit scalar limbs and
  covers keypair; the 4 KiB wipe missed copies 4.7 KiB down under AArch64
  UBSan, so the four secret entry points wipe 8 KiB via an internal
  ama_stack_wipe_below(). The public 4 KiB wipe is unchanged.
- Ascon-AEAD128 had no wipe, though the header claimed every AEAD entry point
  had one; the shipped .so left key words in the entry frame. The bodies now
  run in noinline workers with a wipe after them. The AEAD probe runs each
  call below a 512-byte gap, and both probes link the shared library (the
  LTO test archive hid the leak).
- FROST: reject non-canonical z_i (z+L passed share verification and
  aggregation); admit and attribute every commitment before R is built
  (an undecodable one aborted anonymously); compute binding factors once.
- AMA_DISPATCH_ONLY: sha3-avx512x4, the SVE2 pins and chacha20-neon resolve
  on the probed tier, not on a slot auto-tune or an opt-out may have
  cleared. The unreachable setuid diagnostic is removed.
- -Wsign-conversion is enabled in every strict CI lane after fixing its
  reports at source; __int128 is declared with __extension__ and the
  int128-extension allowlist entry is deleted; dead NEON statics and an
  unused x25519 static are removed.
- INVARIANT-13: the no_sanitize_address exemption is retracted, since ASan
  passes without it. The gate register is removed and the suppression
  regex widened.
- update_docs --counts regenerates C-suite and inventory counts.
- Ed25519 lows: one wNAF recoder, dead fe51 helpers deleted, verify/batch
  @return docs corrected, stale comments fixed.

Every new test fails against the code it replaces (AGENTS.md §6.2).

+++ tests/c/test_nistp.c
@@ -268,7 +268,7 @@ static void test_public_api(void) {
         unsigned i;
 
         for (i = 0; i < sizeof(digest); i++) {
-            digest[i] = (uint8_t)(i * 7u + idx);
+            digest[i] = (uint8_t)(i * 7u + (unsigned)idx);
         }
 
         CHECK(ama_nistp_keypair(curve, priv, pub) == AMA_SUCCESS, "keypair failed");
